// src/types/venue-info.ts
export type VenueInfoLocationOption = {
  id: string
  label: string
}

export type VenueInfoRecord = {
  locationId: string
  locationLabel: string
  venueName: string
  shortName: string
  address1: string
  address2: string
  city: string
  stateProvince: string
  postalCode: string
  phone: string
  extension: string
  phoneTextAlternative: string
  fax: string
}